Creation of "snake.js" and "apple.js"
I've adapted my code to do one of the things the spec sheet asks for.

I created two new classes, "snake.js" which contains all the functions that work with the snake (move the snake, eat the apples for example).

And "apple.js" which contains all the functions that work with the apples. (spawn the apples in random position, draw them for example)